wallet-rpc: Implement HTTP Basic authentication
Author: iamamyth

Digest authentication is ill-suited to a browser context: Most scripts
cannot maintain an accurate per-connection request counter, because the
browser may dispatch a request to an arbitrary, unidentifiable
connection. Furthermore, the form of digest authentication supported
by the RPC server does not validate request bodies, making it
vulnerable to man-in-the-middle attacks, and therefore providing
little benefit over Basic authentication in a number of common
operational environments (such as localhost-only connections,
or SSL-secured connections).